WIP splitting the namespaces with the genesis contract
