+++ Makefile
@@ -7,10 +7,12 @@ SOLC=./node_modules/.bin/solcjs
 TSC=./node_modules/.bin/tsc
 VOTING_CONTRACTS=$(wildcard contracts/*.sol)
 ENS_CONTRACTS=$(wildcard contracts/vendor/registry/*.sol contracts/vendor/resolver/*.sol)
+
 ENS_REGISTRY_ARTIFACT_NAME=contracts_vendor_registry_ENSRegistry_sol_ENSRegistry
-ENS_PUBLIC_RESOLVER_ARTIFACT_NAME=contracts_vendor_resolver_PublicResolver_sol_PublicResolver
-PROCESS_ARTIFACT_NAME=contracts_processes_sol_Processes
-NAMESPACE_ARTIFACT_NAME=contracts_namespaces_sol_Namespaces
+ENS_RESOLVER_ARTIFACT_NAME=contracts_vendor_resolver_PublicResolver_sol_PublicResolver
+PROCESSES_ARTIFACT_NAME=contracts_processes_sol_Processes
+GENESIS_ARTIFACT_NAME=contracts_genesis_sol_Genesis
+NAMESPACES_ARTIFACT_NAME=contracts_namespaces_sol_Namespaces
 TOKEN_STORAGE_PROOF_ARTIFACT_NAME=contracts_token-storage-proof_sol_TokenStorageProof
 TOKEN_STORAGE_PROOF_TEST_ARTIFACT_NAME=contracts_token-storage-proof-test_sol_TokenStorageProofTest
 
@@ -53,7 +55,7 @@ node_modules: package.json
 	@touch $@
 
 js: build/index.js
-contract-output: build/ens-registry.json build/ens-public-resolver.json build/processes.json build/namespaces.json build/token-storage-proof.json build/token-storage-proof-test.json
+contract-output: build/ens-registry.json build/ens-resolver.json build/processes.json build/genesis.json build/namespaces.json build/token-storage-proof.json build/token-storage-proof-test.json
 
 build:
 	@mkdir -p build
@@ -69,20 +71,25 @@ build/ens-registry.json: build/solc/$(ENS_REGISTRY_ARTIFACT_NAME).abi build/solc
 	@echo "Building $@"
 	echo "{\"abi\":$$(cat build/solc/$(ENS_REGISTRY_ARTIFACT_NAME).abi),\"bytecode\":\"0x$$(cat build/solc/$(ENS_REGISTRY_ARTIFACT_NAME).bin)\"}" > $@
 
-build/ens-public-resolver.json: build/solc/$(ENS_PUBLIC_RESOLVER_ARTIFACT_NAME).abi build/solc/$(ENS_PUBLIC_RESOLVER_ARTIFACT_NAME).bin
+build/ens-resolver.json: build/solc/$(ENS_RESOLVER_ARTIFACT_NAME).abi build/solc/$(ENS_RESOLVER_ARTIFACT_NAME).bin
 	@stat $^ > /dev/null
 	@echo "Building $@"
-	echo "{\"abi\":$$(cat build/solc/$(ENS_PUBLIC_RESOLVER_ARTIFACT_NAME).abi),\"bytecode\":\"0x$$(cat build/solc/$(ENS_PUBLIC_RESOLVER_ARTIFACT_NAME).bin)\"}" > $@
+	echo "{\"abi\":$$(cat build/solc/$(ENS_RESOLVER_ARTIFACT_NAME).abi),\"bytecode\":\"0x$$(cat build/solc/$(ENS_RESOLVER_ARTIFACT_NAME).bin)\"}" > $@
 
-build/processes.json: build/solc/$(PROCESS_ARTIFACT_NAME).abi build/solc/$(PROCESS_ARTIFACT_NAME).bin
+build/processes.json: build/solc/$(PROCESSES_ARTIFACT_NAME).abi build/solc/$(PROCESSES_ARTIFACT_NAME).bin
 	@stat $^ > /dev/null
 	@echo "Building $@"
-	echo "{\"abi\":$$(cat build/solc/$(PROCESS_ARTIFACT_NAME).abi),\"bytecode\":\"0x$$(cat build/solc/$(PROCESS_ARTIFACT_NAME).bin)\"}" > $@
+	echo "{\"abi\":$$(cat build/solc/$(PROCESSES_ARTIFACT_NAME).abi),\"bytecode\":\"0x$$(cat build/solc/$(PROCESSES_ARTIFACT_NAME).bin)\"}" > $@
 
-build/namespaces.json: build/solc/$(NAMESPACE_ARTIFACT_NAME).abi build/solc/$(NAMESPACE_ARTIFACT_NAME).bin
+build/genesis.json: build/solc/$(GENESIS_ARTIFACT_NAME).abi build/solc/$(GENESIS_ARTIFACT_NAME).bin
 	@stat $^ > /dev/null
 	@echo "Building $@"
-	echo "{\"abi\":$$(cat build/solc/$(NAMESPACE_ARTIFACT_NAME).abi),\"bytecode\":\"0x$$(cat build/solc/$(NAMESPACE_ARTIFACT_NAME).bin)\"}" > $@
+	echo "{\"abi\":$$(cat build/solc/$(GENESIS_ARTIFACT_NAME).abi),\"bytecode\":\"0x$$(cat build/solc/$(GENESIS_ARTIFACT_NAME).bin)\"}" > $@
+
+build/namespaces.json: build/solc/$(NAMESPACES_ARTIFACT_NAME).abi build/solc/$(NAMESPACES_ARTIFACT_NAME).bin
+	@stat $^ > /dev/null
+	@echo "Building $@"
+	echo "{\"abi\":$$(cat build/solc/$(NAMESPACES_ARTIFACT_NAME).abi),\"bytecode\":\"0x$$(cat build/solc/$(NAMESPACES_ARTIFACT_NAME).bin)\"}" > $@
 
 build/token-storage-proof.json: build/solc/$(TOKEN_STORAGE_PROOF_ARTIFACT_NAME).abi build/solc/$(TOKEN_STORAGE_PROOF_ARTIFACT_NAME).bin
 	@stat $^ > /dev/null
@@ -96,12 +103,14 @@ build/token-storage-proof-test.json: build/solc/$(TOKEN_STORAGE_PROOF_TEST_ARTIF
 
 build/solc/$(ENS_REGISTRY_ARTIFACT_NAME).abi: build/solc
 build/solc/$(ENS_REGISTRY_ARTIFACT_NAME).bin: build/solc
-build/solc/$(ENS_PUBLIC_RESOLVER_ARTIFACT_NAME).abi: build/solc
-build/solc/$(ENS_PUBLIC_RESOLVER_ARTIFACT_NAME).bin: build/solc
-build/solc/$(PROCESS_ARTIFACT_NAME).abi: build/solc
-build/solc/$(PROCESS_ARTIFACT_NAME).bin: build/solc
-build/solc/$(NAMESPACE_ARTIFACT_NAME).abi: build/solc
-build/solc/$(NAMESPACE_ARTIFACT_NAME).bin: build/solc
+build/solc/$(ENS_RESOLVER_ARTIFACT_NAME).abi: build/solc
+build/solc/$(ENS_RESOLVER_ARTIFACT_NAME).bin: build/solc
+build/solc/$(PROCESSES_ARTIFACT_NAME).abi: build/solc
+build/solc/$(PROCESSES_ARTIFACT_NAME).bin: build/solc
+build/solc/$(GENESIS_ARTIFACT_NAME).abi: build/solc
+build/solc/$(GENESIS_ARTIFACT_NAME).bin: build/solc
+build/solc/$(NAMESPACES_ARTIFACT_NAME).abi: build/solc
+build/solc/$(NAMESPACES_ARTIFACT_NAME).bin: build/solc
 build/solc/$(STORAGE_PROOF_ARTIFACT_NAME).abi: build/solc
 build/solc/$(STORAGE_PROOF_ARTIFACT_NAME).bin: build/solc
 build/solc/$(STORAGE_PROOF_TEST_ARTIFACT_NAME).abi: build/solc
